+++ custom-scripts/run-ee-minimal.sh
@@ -1,10 +1,25 @@
 #!/bin/bash
 
 for i in `ls -1 ./artifacts/*.ear`; do
-	./rhamt-mark/bin/rhamt-cli --input "$i" --output "./reports-ee-minimal/${i/.\/artifacts\//}" --archives . --source duo --target duo --packages nl --exportCSV --keepWorkDirs --batchMode --overwrite --userRulesDirectory ./rules-ee-minimal
 
-	sudo find /tmp -amin +10 -exec rm -rf {} +
+	if [ ! -s "./reports-ee-minimal/${i/.\/artifacts\//}/AllIssues.csv" ]; then
+
+		echo "Running analyzer to create report for: ${i/.\/artifacts\//}"
+
+		until ./rhamt-mark/bin/rhamt-cli --input "$i" --output "./reports-ee-minimal/${i/.\/artifacts\//}" --archives . --source duo --target duo --packages nl --exportCSV --keepWorkDirs --batchMode --overwrite --userRulesDirectory ./rules-ee-minimal
+		do
+			echo "An error occured while analyzing ${i/.\/artifacts\//}"
+			sleep 5
+		done
+		
+		sudo find /tmp -amin +10 -exec rm -rf {} +
+
+	else
+		echo "Report already created, skipping: ${i/.\/artifacts\//}"
+	fi
 done
+
+sudo chmod -R 777 reports-ee-minimal
 
 find reports-ee-minimal -name 'AllIssues.csv' | xargs cat > report-ee-minimal.csv
